Refactored modules - separate common dir for common functionality. - separate CLI for client to allow for lightweight clients. - separate CLI for reducer, combiner and monitor.
