+++ fedn/cli/run_cmd.py
@@ -5,44 +5,6 @@ from deprecated import deprecated
 
 from .main import main
 
-@deprecated(version='0.2.0', reason="You should switch to using a configuration file.")
-def get_statestore_config_from_env():
-    import os
-    config = {
-        'type': 'MongoDB',
-        'mongo_config': {
-            'username': os.environ.get('FEDN_MONGO_USER', 'default'),
-            'password': os.environ.get('FEDN_MONGO_PASSWORD', 'password'),
-            'host': os.environ.get('FEDN_MONGO_HOST', 'localhost'),
-            'port': int(os.environ.get('FEDN_MONGO_PORT', '27017')),
-        },
-    }
-    return config
-
-@deprecated(version='0.2.0', reason="You should switch to using a configuration file.")
-def get_storage_config_from_env():
-    """ Deprecated. """
-    import os
-    s3_config = {
-        'storage_type': 'S3',
-        'storage_config': {
-            'storage_access_key': os.environ['FEDN_MINIO_ACCESS_KEY'],
-            'storage_secret_key': os.environ['FEDN_MINIO_SECRET_KEY'],
-            'storage_bucket': 'fednmodels',
-            'storage_secure_mode': False,
-            'storage_hostname': os.environ['FEDN_MINIO_HOST'],
-            'storage_port': int(os.environ['FEDN_MINIO_PORT'])
-            },
-        }
-    return s3_config
-
-@deprecated(version='0.2.0', reason="You should switch to using a configuration file.")
-def get_network_id_from_env():
-    """ Deprecated. """
-    import os
-    network_id = os.environ['ALLIANCE_UID']
-    return network_id
-
 
 def get_statestore_config_from_file(init):
     with open(init, 'r') as file:
@@ -59,46 +21,6 @@ def run_cmd(ctx):
     # if daemon:
     #    print('{} NYI should run as daemon...'.format(__file__))
     pass
-
-
-@run_cmd.command('client')
-@click.option('-d', '--discoverhost', required=False)
-@click.option('-p', '--discoverport', required=False)
-@click.option('-t', '--token', required=False)
-@click.option('-n', '--name', required=False, default=str(uuid.uuid4()))
-@click.option('-i', '--client_id', required=False)
-@click.option('-r', '--remote', required=False, default=True, help='Enable remote configured execution context')
-@click.option('-u', '--dry-run', required=False, default=False)
-@click.option('-s', '--secure', required=False, default=True)
-@click.option('-v', '--preshared-cert', required=False, default=False)
-@click.option('-v', '--verify-cert', required=False, default=False)
-@click.option('-c', '--preferred-combiner', required=False, default=False)
-@click.option('-in', '--init', required=False, default=None, help='Set to a filename to (re)init client from file state.')
-@click.pass_context
-def client_cmd(ctx, discoverhost, discoverport, token, name, client_id, remote, dry_run, secure, preshared_cert,
-               verify_cert,preferred_combiner, init):
-
-    config = {'discover_host': discoverhost, 'discover_port': discoverport, 'token': token, 'name': name,
-              'client_id': client_id, 'remote_compute_context': remote, 'dry_run': dry_run, 'secure': secure,
-              'preshared_cert': preshared_cert, 'verify_cert': verify_cert,'preferred_combiner':preferred_combiner, 'init':init}
-
-    if config['init']:
-        with open(config['init'], 'r') as file:
-            try:
-                settings = dict(yaml.safe_load(file))
-            except yaml.YAMLError as e:
-                print('Failed to read config from settings file, exiting.',flush=True)
-                raise(e)
-
-        # Read/overide settings from config file
-        if 'controller' in settings:
-            reducer_config = settings['controller']
-            for key,val in reducer_config.items():
-                config[key] = val
-
-    from fedn.client import Client
-    client = Client(config)
-    client.run()
 
 
 @run_cmd.command('reducer')
