--- fedn/fedn/common/exceptions.py
@@ -1,3 +1,0 @@
-
-class ModelError(BaseException):
-    pass
