--- fedn/fedn/common/tracer/tracer.py
@@ -1,7 +1,0 @@
-from abc import ABC, abstractmethod
-
-
-class Tracer(ABC):
-    @abstractmethod
-    def report(self, msg):
-        pass
